+++ frontend/.eslintrc.js
@@ -1,0 +1,6 @@
+module.exports = {
+    rules: {
+        '@typescript-eslint/no-unused-vars': 'off',
+        '@typescript-eslint/no-explicit-any': 'off',
+    }
+};
